move overlay content to new folder
